Upload 11/16
I added the Fav Icon to all the pages.
Added image to About Page
Changed size of "About" and "Singles" headlines
I tried to separate the two columns of text on the Tour page, but only made it worse. I was trying to evenly space them apart from one another.
Then the title heading for the Tour page was giving me trouble; i was trying to change the size of "Lady Gaga" and "The Chromatica Ball" separately for different sizes. I was able to change sizes, but they kept sitting next to each other instead of ontop.
I didnt get to changing the size to fit mobile and Ipad yet.